Use geozone names as user segment options
We were depending on translations files for each geozone, which meant
coupling between the database content and the translation files.

Using the names stored in the database solves the problem, even if it
makes the UserSegments class slightly more complex.

Also note now geozone names are the same in every language. This will be
solved when we add database translations to geozones.