Add Security/Open rubocop rule
The `open` method is a bit strange because it can be used to open files
or URLs. In this case, it's clear we're dealing with a URL, so we can
use `URI.parse`.

The code is a bit strange, though, since it returns a value and it has a
side effect, since it opens the URL. I'm not sure about the intention of
the code; if we only wanted to return a value, none of this is needed,
and if we want the test the URL first by opening it, it might be better
to extract that part to a method to show the intention.